python/requirements.txt
numpy
sqlalchemy
pyserial
pandas
autopep8
ipykernel 